requirements.txt
fastapi
uvicorn
requests
python-dotenv
pydantic
psycopg2-binary
passlib[bcrypt]